Выполнение задания 5.14. Модуляция
